requirements.txt
fastapi
uvicorn
langchain-core
langchain-openai
langgraph
tushare
pandas
motor
apscheduler
httpx
python-dotenv
sentence-transformers